test: fix existing test cases after major refactor
