+++ package.json
@@ -10,7 +10,8 @@
     "scripts": {
         "build": "rimraf dist/ && webpack --config webpack.prod.js",
         "test": "jest",
-        "test:watch": "jest --watch",
+        "test:singleWatch": "jest --watch",
+        "test:watch": "jest --watchAll",
         "test:coverage": "jest src --coverage --watchAll --coverageDirectory=\"coverage\"",
         "dev": "webpack --config webpack.dev.js",
         "prepare": "npm run build && husky install",
