+++ src/liveEditor/utils/liveEditorPostMessage.ts
@@ -7,6 +7,7 @@ if (typeof window !== "undefined") {
     liveEditorPostMessage = new EventManager(LIVE_EDITOR_CHANNEL_ID, {
         target: window.parent,
         debug: false,
+        suppressErrors: true,
     });
 }
